+++ Api-Module/src/main/java/com/pawith/apimodule/domain/user/UserAuthorityController.kt
@@ -1,4 +1,4 @@
-package com.pawith.userpresentation
+package com.pawith.apimodule.domain.user
 
 import com.pawith.userapplication.dto.response.UserAuthorityInfoResponse
 import com.pawith.userapplication.service.UserAuthorityGetService
